init for imsqti
