+++ app/functions/src/imsqti.ts
@@ -1,0 +1,26 @@
+import * as functions from "firebase-functions";
+import * as admin from 'firebase-admin';
+import { User } from "./models/user.model";
+import { Test } from "./models/test.model";
+
+export const generateImsqti = functions.https.onCall(async (testId: string, context) => {
+  // check if teacher is signed in
+  if (!context.auth) return;
+  const userId = context.auth.uid
+  const user: User = (await admin.firestore().doc(`users/${userId}`).get()).data() as User
+  if (user.role != 'teacher') throw new functions.https.HttpsError('permission-denied', 'not signed in as teacher')
+
+  // fetch test
+  const test: Test = (await admin.firestore()
+    .doc(`tests/${testId}`)
+    .get())
+    .data() as Test
+  test.id = testId
+
+
+  return getImsQti(test)
+})
+
+function getImsQti(test:Test) {
+  console.log(test)
+}
