+++ app/src/app/services/imsqti.service.ts
@@ -1,0 +1,14 @@
+import { Injectable } from '@angular/core';
+
+@Injectable({
+  providedIn: 'root'
+})
+export class ImsqtiService {
+
+  constructor() { }
+
+  getImsqti() {
+    console.log('get imsqti')
+    
+  }
+}
